requirements.txt
aesedb==0.1.3
aiocmd==0.1.2
aioconsole==0.7.0
aiosmb==0.4.4
aiosqlite==0.21.0
aiowinreg==0.0.7
anyio==4.7.0
arc4==0.3.0
argcomplete==3.5.3
asciitree==0.3.3
asgiref==3.8.1
asn1crypto==1.5.1
asn1tools==0.166.0
asttokens==3.0.0
asyauth==0.0.20
asysocks==0.2.12
attrs==25.1.0
autocommand==2.2.2
autorecon @ git+https://github.com/Tib3rius/AutoRecon.git@457b2048f630f5c5003cd37c78cefb6121a1f7fa
babel==2.17.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
binwalk==2.4.3
bitstruct==8.19.0
blinker==1.9.0
bloodhound==1.8.0
Brlapi==0.8.6
Brotli==1.1.0
capstone==5.0.5
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
dirsearch==0.4.3
distro==1.9.0
dnspython==2.6.1
dploot==3.0.0
dsinternals==1.2.4
dunamai==1.23.0
executing==2.2.0
Flask==3.1.0
fonttools==4.55.3
fs==2.4.16
future==1.0.0
gpg==1.24.2
greenlet==3.1.0
gssapi==1.9.0
h11==0.14.0
h2==4.1.0
hashID==3.1.4
hpack==4.0.0
html5lib-modern==1.2
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.0
idna==3.10
impacket==0.10.0
inflect==7.3.1
invoke==2.2.0
ipython==8.30.0
itsdangerous==2.2.0
jaraco.context==6.0.0
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.1
Jinja2==3.1.5
kali-tweaks==2023.3.2
kiwisolver==1.4.7
ldap3==2.9.1
ldapdomaindump==0.9.4
lightdm-gtk-greeter-settings==1.2.3
louis==3.32.0
lsassy==3.1.11
lxml==5.3.1
lz4==4.4.0+dfsg
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
masky==0.2.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.10
minidump==0.0.21
minikerberos==0.4.4
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.3
msldap==0.5.10
neo4j==5.2.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.10.1
netexec==1.3.0
netifaces==0.11.0
ntlm-auth==1.5.0
numpy==1.26.4
olefile==0.47
onboard==1.4.1
opencv-python==4.11.0.86
oscrypto==1.3.0
packaging==24.2
paramiko==3.5.1
parso==0.8.4
pcapy==0.11.5.dev0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
ply==3.11
poetry-dynamic-versioning==1.4.1
pooch==1.8.2
prompt_toolkit==3.0.50
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.1
pyasyncore==1.0.2
pycairo==1.26.1
pycparser==2.22
pycryptodomex==3.20.0
pycurl==7.45.4
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
pylnk3==0.4.2
PyNaCl==1.5.0
pyNfsClient==0.1.5
PyOpenGL==3.1.7
pyOpenSSL==25.0.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypsrp==0.8.1
pypykatz==0.6.10
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
PySocks==1.7.1
pyspnego==0.10.2
python-apt==2.9.8
python-dateutil==2.9.0
python-dotenv==1.0.1
python-libnmap==0.7.3
python-magic==0.4.27
pytz==2024.1
pywerview==0.3.3
pyxdg==0.28
PyYAML==6.0.2
pyzbar==0.1.9
requests==2.32.3
requests_ntlm==1.3.0
rich==13.9.4
scapy==2.6.1
scipy==1.14.1
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
smbmap==1.10.5
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
types-aiofiles==24.1
types-antlr4-python3-runtime==4.13
types-assertpy==1.1
types-atheris==2.3
types-aws-xray-sdk==2.14
types-beautifulsoup4==4.12
types-bleach==6.2
types-boltons==24.1
types-braintree==4.31
types-cachetools==5.5
types-caldav==1.3
types-capturer==3.0
types-cffi==1.16
types-chevron==0.14
types-click-default-group==1.2
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.8
types-corus==0.10
types-croniter==5.0.1
types-dateparser==1.2
types-decorator==5.1
types-defusedxml==0.7
types-Deprecated==1.2.15
types-docker==7.1
types-dockerfile-parse==2.0
types-docutils==0.21
types-editdistance==0.8
types-entrypoints==0.4
types-ExifRead==3.0
types-fanstatic==1.4
types-first==2.0
types-flake8==7.1
types-flake8-bugbear==24.12.12
types-flake8-builtins==2.5
types-flake8-docstrings==1.7
types-flake8-rst-docstrings==0.3
types-flake8-simplify==0.21
types-flake8-typing-imports==1.16
types-Flask-Cors==5.0
types-Flask-Migrate==4.0
types-Flask-SocketIO==5.4
types-fpdf2==2.8.2
types-gdb==15.0
types-gevent==24.11
types-google-cloud-ndb==2.3
types-greenlet==3.1
types-hdbcli==2.23
types-html5lib==1.1
types-httplib2==0.22
types-humanfriendly==10.0
types-hvac==2.3
types-ibm-db==3.2.4
types-icalendar==6.1
types-influxdb-client==1.45
types-inifile==0.4
types-JACK-Client==0.5
types-Jetson.GPIO==2.1
types-jmespath==1.0
types-jsonschema==4.23
types-jwcrypto==1.5
types-keyboard==0.13
types-ldap3==2.9
types-libsass==0.23
types-lupa==2.2
types-lzstring==1.0
types-m3u8==6.0
types-Markdown==3.7
types-mock==5.1
types-mypy-extensions==1.0
types-mysqlclient==2.2
types-nanoid==2.0.0
types-netaddr==1.3
types-netifaces==0.11
types-networkx==3.4.2
types-oauthlib==3.2
types-objgraph==3.6
types-olefile==0.47
types-openpyxl==3.1.5
types-opentracing==2.4
types-paramiko==3.5
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.17.8
types-pep8-naming==0.14
types-pexpect==4.9
types-pika-ts==1.3
types-polib==1.2
types-portpicker==1.6
types-protobuf==5.29.1
types-psutil==6.1
types-psycopg2==2.9.10
types-pyasn1==0.6
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycocotools==2.0
types-pycurl==7.45.4
types-pyfarmhash==0.4
types-pyflakes==3.2
types-pygit2==1.15
types-Pygments==2.18
types-pyinstaller==6.11
types-pyjks==20.0
types-PyMySQL==1.1
types-pynput==1.7.7
types-pyOpenSSL==24.1
types-pyRFC3339==2.0.1
types-PyScreeze==1.0.1
types-pyserial==3.5
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==3.2
types-python-datemath==3.0.1
types-python-dateutil==2.9
types-python-http-client==3.3.7
types-python-jenkins==1.8
types-python-jose==3.3
types-python-nmap==0.7
types-python-xlib==0.33
types-pytz==2024.2
types-pywin32==308
types-pyxdg==0.28
types-PyYAML==6.0
types-qrbill==1.1
types-qrcode==8.0
types-regex==2024.11.6
types-reportlab==4.2.5
types-requests==2.32
types-requests-oauthlib==2.0
types-retry==0.9
types-RPi.GPIO==0.7
types-s2clientprotocol==5
types-seaborn==0.13.2
types-Send2Trash==1.8
types-setuptools==75.6
types-shapely==2.0
types-simplejson==3.19
types-singledispatch==4.1
types-six==1.17
types-slumber==0.7
types-str2bool==1.1
types-tabulate==0.9
types-tensorflow==2.18.0
types-TgCrypto==1.2
types-toml==0.10
types-toposort==1.10
types-tqdm==4.67
types-translationstring==1.4
types-tree-sitter-languages==1.10
types-ttkthemes==3.2
types-ujson==5.10
types-unidiff==0.7
types-untangle==1.2
types-usersettings==1.1
types-uWSGI==2.0
types-vobject==0.9.9
types-waitress==3.0.1
types-WebOb==1.8
types-whatthepatch==1.0
types-workalendar==17.0
types-WTForms==3.2.1
types-wurlitzer==3.1
types-xdgenvpy==3.0
types-xmltodict==0.14
types-zstd==1.5
types-zxcvbn==4.4
typing_extensions==4.12.2
ufoLib2==0.17.0
unicodedata2==15.1.0
unicrypto==0.0.10
Unidecode==1.4.0
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wfuzz==3.1.0
wheel==0.45.1
winacl==0.1.9
xdg==5
xmltodict==0.13.0
zipp==3.21.0
